# repo/platform/contracts/openapi/bff.yaml
openapi: 3.0.3
info:
  title: OneLink BFF
  description: |
    BFF 聚合层（App/Web 面向接口）— **Phase 2 冻结版**。

    **已实现端点**：
    - `POST /api/v1/bff/auth/register` — 注册代理（identity-service）
    - `POST /api/v1/bff/auth/login` — 登录代理（identity-service）
    - `GET /api/v1/bff/chat/init` — 聊天页初始化
    - `POST /api/v1/bff/chat/messages` — 聊天消息发送代理（ai-chat-service）
    - `GET /api/v1/bff/onboarding` — 冷启动问卷聚合
    - `GET /api/v1/bff/home` — 首页聚合
    - `GET /api/v1/bff/profile/{userId}` — 用户画像
    - `PATCH /api/v1/bff/profile/me` — 用户画像编辑（代理至 profile-service）
    - `POST /api/v1/bff/find/intent` — 找人意图壳（Phase 2 shell，写入意图后返回 202）
    - `POST /api/v1/bff/analytics/events` — 客户端埋点事件接收（fire-and-forget，返回 202）
    - `POST /api/v1/bff/questions/answers` — 问卷回答代理（question-service）

    **占位端点（Phase 3 实现）**：
    - ~~`GET /api/v1/bff/find/results`~~ — 已实现（Phase 3）
    - ~~`GET /api/v1/bff/dm/list`~~ — 已实现（Phase 3）

    **Phase 3 新增端点**：
    - `GET /api/v1/bff/find/results` — 找人结果（代理至 match-service）
    - `GET /api/v1/bff/dm/list` — 私信列表（代理至 dm-service）
    - `POST /api/v1/bff/dm/send` — 发送私信（含安全审查）
    - `POST /api/v1/bff/safety/report` — 举报用户
    - `POST /api/v1/bff/safety/block` — 拉黑用户

    BFF **不解析** token 语义，仅 **透传** `Authorization` 至下游服务。
    默认端口 **8083**。

    ## 鉴权
    所有 `/api/v1/bff/*` 端点需要 Bearer token（通过 `Authorization` 头透传至 identity-service），
    除 `/api/v1/bff/auth/register` 和 `/api/v1/bff/auth/login` 外。

    ## 降级策略
    - question-service 不可达时，`pending_questions` 降级为 `[]`
...

...
servers:
...
paths:
...
  /api/v1/bff/auth/login:
    post:
      summary: 用户登录代理
      operationId: bffAuthLogin
      tags:
        - Auth
      description: |
        代理至 identity-service POST /api/v1/identity/login。
        不需要 Bearer auth。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 参数错误
        '401':
          description: 凭证无效
        '502':
          description: identity-service 不可用
...
  /api/v1/bff/settings/locale:
    get:
      summary: 可用区域/语言列表
      operationId: bffSettingsLocale
      tags:
        - Settings
      security:
        - bearerAuth: []
      description: A4 端点。返回可用 locale、region、timezone 列表。
      responses:
        '200':
          description: 区域和语言注册表
  /api/v1/bff/compliance/summary:
    get:
      summary: 数据权利摘要
      operationId: bffComplianceSummary
      tags:
        - Compliance
      security:
        - bearerAuth: []
      description: A4 端点。返回用户数据权利及进行中请求摘要。
      responses:
        '200':
          description: 合规摘要
        '401':
          description: 未认证
  /api/v1/bff/compliance/export:
    post:
      summary: 请求数据导出
      operationId: bffComplianceExport
      tags:
        - Compliance
      security:
        - bearerAuth: []
      description: A4 端点。提交数据导出请求。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                export_format:
                  type: string
                  enum: [json, csv]
      responses:
        '200':
          description: 导出请求已提交
        '401':
          description: 未认证
  /api/v1/bff/compliance/delete:
    post:
      summary: 请求数据删除
      operationId: bffComplianceDelete
      tags:
        - Compliance
      security:
        - bearerAuth: []
      description: A4 端点。提交数据删除请求。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scope:
                  type: string
      responses:
        '200':
          description: 删除请求已提交
        '401':
          description: 未认证
  /api/v1/bff/compliance/correction:
    post:
      summary: 请求数据更正
      operationId: bffComplianceCorrection
      tags:
        - Compliance
      security:
        - bearerAuth: []
      description: A4 端点。提交数据更正请求。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                field_name:
                  type: string
      responses:
        '200':
          description: 更正请求已提交
        '401':
          description: 未认证
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    AuthLoginRequest:
      type: object
      required:
        - provider
      properties:
        provider:
          type: string
        email:
          type: string
        phone:
          type: string
        password_hash:
          type: string
        oauth_token:
          type: string
        verification_code:
          type: string
    AuthResponse:
      type: object
      required:
        - user_id
        - session
      properties:
        user_id:
          type: string
          format: uuid
        session:
          type: object
          required:
            - token
            - expires_at
          properties:
            token:
              type: string
            expires_at:
              type: string
              format: date-time
...